requirements.txt
Flask
Flask-Bcrypt
Flask-DebugToolbar
Flask-SQLAlchemy
Flask-WTF
ipython
psycopg2-binary
python-dotenv
